\name{Physical_Constants-package}
\alias{Physical_Constants-package}
\alias{Physical_Constants}
\docType{package}
\title{
Physical_Constants: Loads physical constants from NIST database of Fundamental Physical Constants
}
\description{
Loads physical constants from NIST database Fundamental Physical Constants and assigns them human-readable names
}
\details{
\tabular{ll}{
Package: \tab Physical_Constants\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-06\cr
License: \tab Creative Commons Attribution 3.0 Unported License \cr
}
Loading this package assigns human-readable constant names to fundamental physical constants from a NIST database. These constants are then available for use in the R computing environment.
}
\author{
Pascal Mickelson

Maintainer: Pascal Mickelson, https://github.com/mickelsp/Physical_Constants
}
\references{
Where the data comes from: http://physics.nist.gov/cuu/Constants/index.html
Citation lookup for the source of the data: http://physics.nist.gov/cuu/Constants/Citations/Search.html
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
None yet.
~~ simple examples of the most important functions ~~
}
